+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/api/statuses/voteInPoll.kt
@@ -1,0 +1,30 @@
+package site.remlit.snowdrop.api.statuses
+
+import com.russhwolf.settings.ExperimentalSettingsApi
+import io.ktor.client.request.header
+import io.ktor.client.request.post
+import io.ktor.client.request.setBody
+import site.remlit.snowdrop.model.ApiResponse
+import site.remlit.snowdrop.model.Status
+import site.remlit.snowdrop.model.request.PollVoteRequest
+import site.remlit.snowdrop.util.config.endOfRequest
+import site.remlit.snowdrop.util.config.httpClient
+import site.remlit.snowdrop.util.getCurrentAccountHost
+import site.remlit.snowdrop.util.getCurrentAccountId
+import site.remlit.snowdrop.util.safeApiRequest
+import site.remlit.snowdrop.util.settings
+
+@OptIn(ExperimentalSettingsApi::class)
+suspend fun voteInPoll(id: String, choices: List<Int>): ApiResponse<Status.Poll> = safeApiRequest {
+	val accountId = getCurrentAccountId()
+	val host = getCurrentAccountHost()
+	val token = settings.getString("account_${accountId}_token", "")
+
+	val req = httpClient.post("https://$host/api/v1/polls/$id/votes") {
+		header("Authorization", "Bearer $token")
+		header("Content-Type", "application/json")
+		setBody(PollVoteRequest(choices))
+	}
+
+	endOfRequest(req)
+}