+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/model/request/PollVoteRequest.kt
@@ -1,0 +1,8 @@
+package site.remlit.snowdrop.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PollVoteRequest(
+	val choices: List<Int>
+)